+++ src/devices/ColorLightAccessory.ts
@@ -68,6 +68,7 @@ export class ColorLightAccessory extends BaseMatterAccessory {
   }
 
   private async handleSetLevel(request: MatterRequests.MoveToLevel): Promise<void> {
+    this.logInfo(`MoveToLevel request: ${JSON.stringify(request)}`)
     const { level } = request
     const brightnessPercent = Math.round((level / 254) * 100)
     this.logInfo(`setting brightness to ${brightnessPercent}%.`)
@@ -75,6 +76,7 @@ export class ColorLightAccessory extends BaseMatterAccessory {
   }
 
   private async handleSetColor(request: MatterRequests.MoveToColor): Promise<void> {
+    this.logInfo(`MoveToColor request: ${JSON.stringify(request)}`)
     const { colorX, colorY, transitionTime } = request
     const xFloat = (colorX / 65535).toFixed(4)
     const yFloat = (colorY / 65535).toFixed(4)
@@ -83,6 +85,7 @@ export class ColorLightAccessory extends BaseMatterAccessory {
   }
 
   private async handleSetHueSaturation(request: MatterRequests.MoveToHueAndSaturation): Promise<void> {
+    this.logInfo(`MoveToHueAndSaturation request: ${JSON.stringify(request)}`)
     const { hue, saturation, transitionTime } = request
     const hueDegrees = Math.round((hue / 254) * 360)
     const saturationPercent = Math.round((saturation / 254) * 100)