+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to `@homebridge-plugins/homebridge-matter` will be documente
 ### Changes
 
 - refactor codebase for class-based devices
+- add logging for all `MatterRequest` events
 
 ## v0.1.6 (2025-10-24)